Making edda installable in the main nix-shell
