+++ data/tools/edda/shell.nix
@@ -1,0 +1,19 @@
+with (import <nixpkgs> {});
+let
+  pythonEnv = python311.withPackages(ps: with ps; [
+    click
+    click-log
+
+    black
+    pytest
+  ]);
+
+in pkgs.mkShell {
+  name = "edda";
+  buildInputs = [
+    pythonEnv
+  ];
+
+  shellHook = ''
+  '';
+}
