+++ data/tools/edda/src/about.py
@@ -1,0 +1,1 @@
+_version='0.1.0'
