+++ data/tools/edda/tests/fixtures/test-db/transition_migrations/0002_test_transition.sql
@@ -1,0 +1,1 @@
+SELECT * FROM edda_migrations;
