+++ shell.nix
@@ -1,11 +1,6 @@
 with (import <nixpkgs> {});
 let
   pythonEnv = python311.withPackages(ps: with ps; [
-    click
-    click-log
-
-    black
-    pytest
   ]);
 
 in pkgs.mkShell {
@@ -14,7 +9,6 @@ in pkgs.mkShell {
     nodejs
     actionlint
     k6
-    pythonEnv
     (import ./data/tools/edda/derivation.nix { lib = lib; python311Packages = python311Packages; })
   ];
